+++ NEWS.md
@@ -1,6 +1,6 @@
 # elasticsearchr 0.1.0.900
 
-* ABC...
+* Fixed data.frame index bug - Elasticsearch Bulk API was failing when data.frame was large (>15mb).
 
 # elasticsearchr 0.1.0